# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * 2PUC Notes Store API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface StoreStats {
  totalResources: number;
  freeResources: number;
  paidResources: number;
  totalSections: number;
  totalOrders: number;
}

export interface Section {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  isVisible: boolean;
  sortOrder: number;
  cardCount: number;
  createdAt: string;
}

export interface SectionInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  isVisible?: boolean;
}

export interface SectionUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  description?: string | null;
  isVisible?: boolean;
  sortOrder?: number;
}

export interface ReorderInput {
  ids: number[];
}

export interface Card {
  id: number;
  sectionId: number;
  /** @nullable */
  sectionName?: string | null;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  subject?: string | null;
  /** @nullable */
  chapter?: string | null;
  /** @nullable */
  semester?: string | null;
  board: string;
  classLevel: string;
  resourceType: string;
  isFree: boolean;
  pricePaise: number;
  /** @nullable */
  discountPricePaise?: number | null;
  /** @nullable */
  telegramLink?: string | null;
  isNew: boolean;
  /** @nullable */
  thumbnailUrl?: string | null;
  previewImageUrls: string[];
  /** @nullable */
  pdfFileKey?: string | null;
  /** @nullable */
  pageCount?: number | null;
  /** @nullable */
  fileSizeKb?: number | null;
  isFeatured: boolean;
  isVisible: boolean;
  isDeleted: boolean;
  sortOrder: number;
  createdAt: string;
}

export interface CardInput {
  sectionId: number;
  /** @minLength 1 */
  title: string;
  description?: string;
  subject?: string;
  chapter?: string;
  semester?: string;
  board?: string;
  classLevel?: string;
  resourceType?: string;
  isFree: boolean;
  pricePaise: number;
  discountPricePaise?: number;
  telegramLink?: string;
  isNew?: boolean;
  thumbnailUrl?: string;
  previewImageUrls?: string[];
  pdfFileKey?: string;
  pageCount?: number;
  fileSizeKb?: number;
  isFeatured?: boolean;
  isVisible?: boolean;
}

export interface CardUpdate {
  sectionId?: number;
  title?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  subject?: string | null;
  /** @nullable */
  chapter?: string | null;
  /** @nullable */
  semester?: string | null;
  board?: string;
  classLevel?: string;
  resourceType?: string;
  isFree?: boolean;
  pricePaise?: number;
  /** @nullable */
  discountPricePaise?: number | null;
  /** @nullable */
  telegramLink?: string | null;
  isNew?: boolean;
  /** @nullable */
  thumbnailUrl?: string | null;
  previewImageUrls?: string[];
  /** @nullable */
  pdfFileKey?: string | null;
  /** @nullable */
  pageCount?: number | null;
  /** @nullable */
  fileSizeKb?: number | null;
  isFeatured?: boolean;
  isVisible?: boolean;
  sortOrder?: number;
}

export interface RazorpayOrderInput {
  cardId: number;
  customerName: string;
  customerPhone: string;
  /** @nullable */
  customerEmail?: string | null;
}

export interface RazorpayOrderResult {
  razorpayOrderId: string;
  orderId: string;
  amountPaise: number;
  currency: string;
  keyId: string;
  cardTitle: string;
  customerName: string;
  customerEmail?: string;
  customerPhone: string;
}

export interface VerifyCheckoutInput {
  razorpayOrderId: string;
  razorpayPaymentId: string;
  razorpaySignature: string;
}

export interface FreeDownloadInput {
  cardId: number;
  customerName: string;
  customerPhone: string;
  /** @nullable */
  customerEmail?: string | null;
}

export interface FreeDownloadResult {
  orderId: string;
  downloadUrl: string;
  filename: string;
}

export interface Order {
  id: string;
  cardId: number;
  cardTitle: string;
  /** @nullable */
  customerId?: number | null;
  customerName: string;
  customerPhone: string;
  /** @nullable */
  customerEmail?: string | null;
  amountPaise: number;
  status: string;
  /** @nullable */
  razorpayOrderId?: string | null;
  /** @nullable */
  paidAt?: string | null;
  createdAt: string;
}

export interface OrderList {
  orders: Order[];
  total: number;
  page: number;
}

export interface AdminLoginInput {
  username: string;
  password: string;
}

export interface AdminLoginResult {
  success: boolean;
  message: string;
  token?: string;
}

export interface AdminSession {
  isAdmin: boolean;
}

export interface AdminStats {
  totalRevenuePaise: number;
  totalOrders: number;
  totalCards: number;
  totalSections: number;
  totalCustomers: number;
  recentOrders: Order[];
}

export interface UploadResult {
  url: string;
  key: string;
}

export type ListCardsParams = {
search?: string;
subject?: string;
resourceType?: string;
isFree?: boolean;
sectionId?: number;
featured?: boolean;
};

export type GetPurchasesParams = {
phone: string;
};

export type AdminListCardsParams = {
sectionId?: number;
};

export type AdminListOrdersParams = {
page?: number;
};