# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: 2PUC Notes Store API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResult"
        "401":
          description: Invalid credentials
...
  /admin/me:
    get:
      operationId: adminMe
      tags: [admin]
      summary: Check admin session
      responses:
        "200":
          description: Admin session valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Not authenticated
...
  /admin/sections:
    get:
      operationId: adminListSections
      tags: [admin]
      summary: List all sections (including hidden)
      responses:
        "200":
          description: All sections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Section"
    post:
      operationId: adminCreateSection
      tags: [admin]
      summary: Create a new section
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SectionInput"
      responses:
        "201":
          description: Created section
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Section"
  /admin/sections/{id}:
    patch:
      operationId: adminUpdateSection
      tags: [admin]
      summary: Update a section
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SectionUpdate"
      responses:
        "200":
          description: Updated section
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Section"
    delete:
      operationId: adminDeleteSection
      tags: [admin]
      summary: Delete a section
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /admin/sections/reorder:
    post:
      operationId: adminReorderSections
      tags: [admin]
      summary: Reorder sections
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderInput"
      responses:
        "200":
          description: OK
  /admin/cards:
    get:
      operationId: adminListCards
      tags: [admin]
      summary: List all non-deleted cards (admin view)
      parameters:
        - name: sectionId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: All cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Card"
    post:
      operationId: adminCreateCard
      tags: [admin]
      summary: Create a new card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardInput"
      responses:
        "201":
          description: Created card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
  /admin/cards/{id}:
    patch:
      operationId: adminUpdateCard
      tags: [admin]
      summary: Update a card
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardUpdate"
      responses:
        "200":
          description: Updated card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
    delete:
      operationId: adminDeleteCard
      tags: [admin]
      summary: Soft-delete a card (sets is_deleted=true, removes PDF from storage)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /admin/orders:
    get:
      operationId: adminListOrders
      tags: [admin]
      summary: List all orders (paginated)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderList"
components:
  schemas:
...
    Section:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        isVisible:
          type: boolean
        sortOrder:
          type: integer
        cardCount:
          type: integer
        createdAt:
          type: string
      required: [id, name, slug, isVisible, sortOrder, cardCount, createdAt]
    SectionInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        isVisible:
          type: boolean
      required: [name]
    SectionUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        isVisible:
          type: boolean
        sortOrder:
          type: integer
    ReorderInput:
      type: object
      properties:
        ids:
          type: array
          items:
            type: integer
      required: [ids]
    Card:
      type: object
      properties:
        id:
          type: integer
        sectionId:
          type: integer
        sectionName:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        subject:
          type: ["string", "null"]
        chapter:
          type: ["string", "null"]
        semester:
          type: ["string", "null"]
        board:
          type: string
        classLevel:
          type: string
        resourceType:
          type: string
        isFree:
          type: boolean
        pricePaise:
          type: integer
        discountPricePaise:
          type: ["integer", "null"]
        telegramLink:
          type: ["string", "null"]
        isNew:
          type: boolean
        thumbnailUrl:
          type: ["string", "null"]
        previewImageUrls:
          type: array
          items:
            type: string
        pdfFileKey:
          type: ["string", "null"]
        pageCount:
          type: ["integer", "null"]
        fileSizeKb:
          type: ["integer", "null"]
        isFeatured:
          type: boolean
        isVisible:
          type: boolean
        isDeleted:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
      required: [id, sectionId, title, board, classLevel, resourceType, isFree, pricePaise, isNew, isFeatured, isVisible, isDeleted, sortOrder, previewImageUrls, createdAt]
    CardInput:
      type: object
      properties:
        sectionId:
          type: integer
        title:
          type: string
          minLength: 1
        description:
          type: string
        subject:
          type: string
        chapter:
          type: string
        semester:
          type: string
        board:
          type: string
        classLevel:
          type: string
        resourceType:
          type: string
        isFree:
          type: boolean
        pricePaise:
          type: integer
        discountPricePaise:
          type: integer
        telegramLink:
          type: string
        isNew:
          type: boolean
        thumbnailUrl:
          type: string
        previewImageUrls:
          type: array
          items:
            type: string
        pdfFileKey:
          type: string
        pageCount:
          type: integer
        fileSizeKb:
          type: integer
        isFeatured:
          type: boolean
        isVisible:
          type: boolean
      required: [sectionId, title, isFree, pricePaise]
    CardUpdate:
      type: object
      properties:
        sectionId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        subject:
          type: ["string", "null"]
        chapter:
          type: ["string", "null"]
        semester:
          type: ["string", "null"]
        board:
          type: string
        classLevel:
          type: string
        resourceType:
          type: string
        isFree:
          type: boolean
        pricePaise:
          type: integer
        discountPricePaise:
          type: ["integer", "null"]
        telegramLink:
          type: ["string", "null"]
        isNew:
          type: boolean
        thumbnailUrl:
          type: ["string", "null"]
        previewImageUrls:
          type: array
          items:
            type: string
        pdfFileKey:
          type: ["string", "null"]
        pageCount:
          type: ["integer", "null"]
        fileSizeKb:
          type: ["integer", "null"]
        isFeatured:
          type: boolean
        isVisible:
          type: boolean
        sortOrder:
          type: integer
...
    Order:
      type: object
      properties:
        id:
          type: string
        cardId:
          type: integer
        cardTitle:
          type: string
        customerId:
          type: ["integer", "null"]
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: ["string", "null"]
        amountPaise:
          type: integer
        status:
          type: string
        razorpayOrderId:
          type: ["string", "null"]
        paidAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, cardId, cardTitle, customerName, customerPhone, amountPaise, status, createdAt]
    OrderList:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
      required: [orders, total, page]
    AdminLoginInput:
...