# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * 2PUC Notes Store API
 * OpenAPI spec version: 0.2.0
 */

export interface Order {
  id: string;
  cardId: number;
  cardTitle: string;
  /** @nullable */
  customerId?: number | null;
  customerName: string;
  customerPhone: string;
  /** @nullable */
  customerEmail?: string | null;
  amountPaise: number;
  status: string;
  /** @nullable */
  razorpayOrderId?: string | null;
  /** @nullable */
  paidAt?: string | null;
  createdAt: string;
}