# lib/api-zod/src/generated/types/orderList.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * 2PUC Notes Store API
 * OpenAPI spec version: 0.2.0
 */
import type { Order } from './order';

export interface OrderList {
  orders: Order[];
  total: number;
  page: number;
}